LibLSLCC.LibraryData.Reflection attributes added.
Added explicit attributes to all optional/core modules
so that LSLLibraryDataReflectionSerializer does not need
to specify type converters in the Compiler class.

No type converters are defined in the Compiler class now,
so if a module does not explicitly attribute the parameters
of a property/field/method marked with [LSLFunctionAttribute]/[LSLConstantAttribute] or provide a class wide/member specific converter for it then an exception
will be generated when loading the module with the LibLSLCC
compiler assembly loaded.

Modified IScriptModuleComms and ScriptModuleCommsModule to provide
more information about the constants and functions defined in
modules that register script invocations/constants.

The old compiler had slight changes in two places to accept the
new interface, and still works just the same when not using the
LibLSLCC compiler.