Fixed a ScriptModuleCommsModule mistake.
was searching for 'ScriptConstantAttribute' instead of
'ScriptInvocationAttribute' inside RegisterScriptInvocations.

This was not manifesting a problem since RegisterScriptInvocations
also looks for 'LSLFunctionAttribute'.

I have attributed all of the modules that come with OpenSim
using 'LSLFunctionAttribute' from LibLSLCC, in addition to their
existing 'ScriptInvocationsAttribute's.

If I did not look at the diff I would not have found this for a while.