Critical bug fix for component access on vec/rot
where:

vector/rotation some_var = <0,0,0>;

this:

some_var.x = 2;

was being optimized away.  It was because the expression property that
looks
down the tree for modifiable L-Value's was failing to recognize that
the component access '.' operator, constitutes a modifiable value that
can affect program state.

Update from devel, see commit:
https://github.com/EriHoss/LibLSLCC/commit/e854acc2de1b670b98f60fa6757e7774beaee3e6